+++ src/lex/token.rs
@@ -13,9 +13,9 @@ impl Lex {
     pub fn new(pos: &CaretPos, token: Token) -> Self {
         let start = pos.clone();
         let end = if let Token::Str(_str, _) = &token {
-            pos.clone().offset_line(max(_str.lines().count() - 1, 0))
+            pos.clone().offset_line(max((_str.lines().count() as i32 - 1) as usize, 0))
         } else if let Token::DocStr(_str) = &token {
-            pos.clone().offset_line(max(_str.lines().count() - 1, 0))
+            pos.clone().offset_line(max((_str.lines().count() as i32 - 1) as usize, 0))
         } else {
             pos.clone()
         };
